requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-js-asset==2.2.0
frozenlist==1.4.1
geoip2==4.8.0
idna==3.10
maxminddb==2.6.2
multidict==6.1.0
pillow==10.4.0
psycopg2==2.9.9
pytz==2024.2
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.3
yarl==1.11.1
